fix: 게임을 실행하는 게임기 클래스를 만들어 구현하였다. - 게임이 그대로 시작하는게 아니라, 따로 게임기의 기능을 갖는 GameEmulator 클래스를 만들어서 분리하였다.
